requirements.txt
# python version 3.8
discord.py==2.3.2
transformers~=4.39.3
torch~=1.10.1
tensorboardX~=2.6.2.2
datasets~=2.18.0
tqdm~=4.66.1
evaluate~=0.4.1
numpy~=1.24.4
nltk~=3.8.1
matplotlib~=3.5.1
scikit-learn~=1.0.2
tensorflow~=2.13.0